src/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
azure-core==1.32.0
azure-storage-blob==12.24.0
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
comtypes==1.4.7
confluent-kafka==2.8.0
contourpy==1.3.0
coverage==7.6.8
cryptography==43.0.1
cycler==0.12.1
distro==1.9.0
Django==5.1.2
django-cors-headers==4.5.0
django-prometheus==2.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.8
email_validator==2.2.0
exceptiongroup==1.2.2
fonttools==4.54.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
isodate==0.7.2
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.6
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-text-splitters==0.3.1
langsmith==0.1.139
lxml==5.3.0
matplotlib==3.9.2
multidict==6.1.0
numpy==1.26.4
openai==1.53.0
opencv-python==4.10.0.84
orjson==3.10.10
packaging==23.2
pdfminer.six==20231228
pillow==10.4.0
prometheus_client==0.21.1
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMuPDF==1.24.13
PyMuPDFb==1.24.10
pyparsing==3.2.0
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.5.2
scikit-metrics==0.1.0
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
tblib==3.0.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.17.1
